requirements.txt
autopep8==1.2.1
BeautifulSoup==3.2.1
coverage==4.0.2
Django==1.8.6
django-debug-toolbar==1.4
django-jenkins==0.18.1
evernote==1.25.1
flake8==2.2.4
flake8-import-order==0.6.1
flake8-pep257==1.0.3
funcsigs==0.4
hacking==0.10.2
httplib2==0.9.2
httpretty==0.8.12
linecache2==1.0.0
mccabe==0.2.1
mock==1.3.0
MySQL-python==1.2.5
oauth2==1.9.0.post1
oauthlib==1.0.3
pbr==1.8.1
pep257==0.7.0
pep8==1.5.7
pyflakes==0.8.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
sqlparse==0.1.18
traceback2==1.4.0
unittest2==1.1.0
wheel==0.24.0
